+++ mtc_api/database/migrations/2020_07_30_143144_create_guides_tourism_types_table.php
@@ -15,8 +15,8 @@ class CreateGuidesTourismTypesTable extends Migration
     {
         Schema::create('guides_tourism_types', function (Blueprint $table) {
             $table->bigIncrements('guides_tourism_type_id');
-            $table->unsignedInteger('tourism_type_id');
-            $table->unsignedInteger('guide_id');
+            $table->unsignedBigInteger('tourism_type_id');
+            $table->unsignedBigInteger('guide_id');
             $table->timestamps();
             $table->softDeletes();
