+++ mtc_api/database/migrations/2020_07_30_155001_add_relation_between_sites_cities_tourism_types_table.php
@@ -14,8 +14,8 @@ class AddRelationBetweenSitesCitiesTourismTypesTable extends Migration
     public function up()
     {
         Schema::table('sites', function (Blueprint $table) {
-            $table->unsignedInteger('citie_id');
-            $table->unsignedInteger('tourism_type_id');
+            $table->unsignedBigInteger('citie_id');
+            $table->unsignedBigInteger('tourism_type_id');
 
             $table->foreign('citie_id')->references('citie_id')->on('cities');
             $table->foreign('tourism_type_id')->references('tourism_type_id')->on('tourism_types');
