+++ mtc_api/database/migrations/2020_07_30_155617_add_relation_between_clients_users_hotels_table.php
@@ -14,8 +14,8 @@ class AddRelationBetweenClientsUsersHotelsTable extends Migration
     public function up()
     {
         Schema::table('clients', function (Blueprint $table) {
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('hotel_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('hotel_id');
 
             $table->foreign('user_id')->references('user_id')->on('users');
             $table->foreign('hotel_id')->references('hotel_id')->on('hotels');
